# i_ringer_was/i_ringer_was/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/signin": {
      "post": {
        "operationId": "AuthController_signinUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "description": "JWT Access Token (2시간 유효)"
                    },
                    "refresh_token": {
                      "type": "string",
                      "description": "JWT Refresh Token (30일 유효)"
                    },
                    "user": {
                      "type": "object",
                      "description": "사용자 정보",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "등록되지 않은 사용자입니다 또는 비밀번호가 일치하지 않습니다"
          }
        },
        "summary": "사용자 로그인",
        "tags": [
          "인증"
        ]
      }
    },
    "/api/user/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string",
                    "description": "리프레시 토큰"
                  }
                },
                "required": [
                  "refresh_token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "토큰 갱신 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "description": "새로운 JWT Access Token (2시간 유효)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "유효하지 않은 리프레시 토큰"
          }
        },
        "summary": "토큰 갱신",
        "tags": [
          "인증"
        ]
      }
    },
...
    "/monitoring": {
      "get": {
        "description": "병원, 병동, 병실별로 환자 및 링거 상태를 모니터링합니다.",
        "operationId": "MonitoringController_getMonitoringData",
        "parameters": [
          {
            "name": "hospital_id",
            "required": false,
            "in": "query",
            "description": "병원 ID (단일값)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "ward_id",
            "required": false,
            "in": "query",
            "description": "병동 ID (단일값)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "room_id",
            "required": false,
            "in": "query",
            "description": "병실 ID (콤마로 구분된 다중값)",
            "schema": {
              "example": "1,2,3",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "모니터링 데이터 조회 성공"
          }
        },
        "summary": "모니터링 데이터 조회",
        "tags": [
          "Monitoring"
        ]
      }
    }
  },
  "info": {
    "title": "I-Ringer API",
    "description": "Hospital IV drip monitoring system API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "기본 서버"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserLoginDto": {
        "type": "object",
        "properties": {
          "auth_id": {
            "type": "string",
            "example": "nurse001",
            "description": "사용자 아이디(auth_id)"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "비밀번호"
          }
        },
        "required": [
          "auth_id",
          "password"
        ]
      },
...
    }
  }
}